backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.28",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "groq>=1.0.0",
    "nltk>=3.9.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "celery>=5.3",
    "redis>=5.0",
    "django-celery-results>=2.5",
    "transformers>=4.57.6",
    "torch>=2.8.0",
    "beautifulsoup4>=4.14.3",
    "spacy>=3.8.11",
    "apify-client>=1.8.1",
]
